Delete the vga.c and text_on_vga.c files from ./lib . Add a temporary GUI to ./system, which unifies deleted files.
