+++ shell/mouse.c
@@ -19,15 +19,12 @@
 int mouse_x=0;         //signed char
 int mouse_y=0;         //signed char
 
-extern void print_text_on_vga(unsigned int x, unsigned int y, const char *text);
-extern void borrar(int x1, int y1, int x2, int y2);
-
 process mouse_p (
                 did32   dev             /* ID of tty device from which  */
         )                               /*   to accept commands         */
 {
 	int old_x=0;
-	int old_y=0;;
+	int old_y=0;
 	unsigned char c;
 	char * h = "Xinu x86 GUI experiments: VGA display, mouse and kbd";
 	print_text_on_vga(10, 500, h);
@@ -49,16 +46,22 @@ process mouse_p (
 		else
 		 	sprintf(texto, "mouse x: %d  y: %d                      ", mouse_x, mouse_y);
 
-
 		pixel(mouse_x, mouse_y, 0x00ff0000);
 		pixel(mouse_x+1, mouse_y, 0x00ff0000);
 		pixel(mouse_x-1, mouse_y, 0x00ff0000);
 		pixel(mouse_x, mouse_y-1, 0x00ff0000);
 		pixel(mouse_x, mouse_y+1, 0x00ff0000);
+		
 		print_text_on_vga(10, 10, texto);
 		old_x = mouse_x; old_y = mouse_y;
 		c = 0;
 		sleepms(1);
+
+		pixel(old_x, old_y, 0x00ffff00);
+		pixel(old_x+1, old_y, 0x00ffff00);
+		pixel(old_x-1, old_y, 0x00ffff00);
+		pixel(old_x, old_y-1, 0x00ffff00);
+		pixel(old_x, old_y+1, 0x00ffff00);
 
 	}
 	close(MOUSE);
