+++ system/main.c
@@ -2,18 +2,12 @@
 
 #include <xinu.h>
 
-
-extern paint_screen();
-extern print_text_on_vga(unsigned int x, unsigned int y, const char * t);
-
 process	main(void)
 {
-
-
 	paint_screen();
 	print_text_on_vga(10, 200, "Xinu OS for PC with VGA support");
 	print_text_on_vga(10, 220, "Sistemas Operativos I");
-
+	
 	recvclr();
 	resume(create(shell, 8192, 50, "shell", 1, CONSOLE));
 	resume(create(mouse_p, 1024, 50, "mouse", 1, CONSOLE));
